+++ src/app/pages/quickMaps/pages/costEffectiveness/interventionReview/pages/interventionCompliance/interventionCompliance.component.ts
@@ -6,7 +6,7 @@ import {
   BaselineAssumptions,
   InterventionBaselineAssumptions,
 } from 'src/app/apiAndObjects/objects/interventionBaselineAssumptions';
-import { FoodVehicleStandard } from 'src/app/apiAndObjects/objects/iiiiiiinterventionFoodVehicleStandards';
+import { FoodVehicleStandard } from 'src/app/apiAndObjects/objects/interventionFoodVehicleStandards';
 import { QuickMapsService } from 'src/app/pages/quickMaps/quickMaps.service';
 import { AppRoutes } from 'src/app/routes/routes';
 import { InterventionDataService } from 'src/app/services/interventionData.service';
