+++ src/app/pages/quickMaps/pages/costEffectiveness/interventionReview/utilities/microNutrientsInPremixTable/premixTableRow/premixTableRow.component.ts
@@ -5,7 +5,7 @@ import { BaselineAssumptions } from 'src/app/apiAndObjects/objects/interventionB
 import {
   FoodVehicleCompound,
   FoodVehicleStandard,
-} from 'src/app/apiAndObjects/objects/iiiiiiinterventionFoodVehicleStandards';
+} from 'src/app/apiAndObjects/objects/interventionFoodVehicleStandards';
 import { InterventionDataService } from 'src/app/services/interventionData.service';
 
 @Component({
